requirements.txt

streamlit==1.26.0
pandas==1.5.3
matplotlib==3.6.2
psycopg2-binary==2.9.5
scikit-learn==1.2.0
xgboost==1.7.2
